fix tx bit holder

+++ Alt9SoftSerial.cpp
@@ -51,7 +51,7 @@ static volatile uint16_t rx_buffer[RX_BUFFER_SIZE];
 
 static volatile uint8_t tx_state=0;
 static uint16_t tx_byte;
-static uint8_t tx_bit;
+static uint16_t tx_bit;
 static volatile uint8_t tx_buffer_head;
 static volatile uint8_t tx_buffer_tail;
 #define TX_BUFFER_SIZE 68
